+++ packages/ranger-ui/src/LineContainer/index.ts
@@ -1,0 +1,2 @@
+export type { LineContainerProps } from './lineContainer'
+export { default as LineContainer } from './lineContainer'