+++ packages/ranger-ui/src/NextPerformance/index.ts
@@ -1,0 +1,2 @@
+export type { NextPerformanceProps } from './nextPerformance'
+export { default as NextPerformance } from './nextPerformance'