+++ packages/ranger-ui/src/typings/emotion.d.ts
@@ -1,0 +1,9 @@
+import type { Theme as EmotionTheme } from '@emotion/react'
+
+declare module '@emotion/react' {
+  export interface Theme extends EmotionTheme {
+    width: {
+      container: string
+    }
+  }
+}